requirements.txt
datasets==3.3.0
numpy==1.21.5
openai==1.63.2
pandas==1.4.2
python-dotenv==1.0.1
PyYAML==6.0.2
tqdm==4.67.1
